move compiler to bundler package
