+++ packages/docusaurus-bundler/package.json
@@ -18,12 +18,14 @@
   },
   "license": "MIT",
   "dependencies": {
+    "@docusaurus/cssnano-preset": "3.5.2",
     "@docusaurus/faster": "3.5.2",
     "@docusaurus/logger": "3.5.2",
     "@docusaurus/types": "3.5.2",
     "copy-webpack-plugin": "^11.0.0",
     "css-minimizer-webpack-plugin": "^5.0.1",
     "mini-css-extract-plugin": "^2.9.1",
+    "react-dev-utils": "^12.0.1",
     "terser-webpack-plugin": "^5.3.9",
     "tslib": "^2.6.0",
     "webpack": "^5.88.1",
